// src/openapi/khronos/apis/PredictionApi.ts
export const GetAllPredictionsFilteredStatusEnum = {
    Pending: 'pending',
    Completed: 'completed',
    Error: 'error'
} as const;
export type GetAllPredictionsFilteredStatusEnum = typeof GetAllPredictionsFilteredStatusEnum[keyof typeof GetAllPredictionsFilteredStatusEnum];